+++ qr.go
@@ -24,7 +24,7 @@ import (
 	"fmt"
 	"net/http"
 
-	whttp "github.com/piusalfred/whatsapp/http"
+	whttp "github.com/piusalfred/whatsapp/pkg/http"
 )
 
 const (
